requirements.txt
# 基础依赖
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# 深度学习框架
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
segmentation-models-pytorch==0.3.3
timm==0.9.7  # 用于DeepLabV3+的预训练骨干网络

# 图像处理库
opencv-python==4.8.0.74
spectral==0.23.1
pillow==10.0.0

# 遥感数据处理库
rasterio==1.3.7
geopandas==0.13.2
GDAL==3.6.4

# 可视化库
matplotlib==3.7.2
seaborn==0.12.2

# 工具库
tqdm==4.65.0
pyyaml==6.0.1

# 评估工具
ttach==0.0.3  # 测试时增强库